chess-vision-fast/backend/requirements.txt
fastapi[all]
uvicorn[standard]
python-multipart
Pillow
numpy
opencv-python
python-chess
ultralytics
pydantic
websockets
pydantic-settings
